Drive by warning fix.

+++ lib/usb/serial.cc
@@ -91,7 +91,7 @@ public:
     void setBaudRate(int baudRate) override
     {
         DCB dcb = {.DCBlength = sizeof(DCB),
-            .BaudRate = baudRate,
+            .BaudRate = (DWORD) baudRate,
             .fBinary = true,
             .ByteSize = 8,
             .Parity = NOPARITY,
